Implement more actions for the piano roll
Includes a lot of other things, like the start of a timeline
implementation, the start of the localization support, the start of a
command palette, the start of user-defined shortcuts, etc.

There's a lot more false starts that were not captured in this commit,
and breaking this up into smaller commits wouldn't really make a
difference - there's probably no reason to sift through that. At least
the build is not broken!