Fix offscreen note indicator for flat notes
That was showing up too early for those. I could have kept the check in
"world space" (i.e. checking pitch values directly) but it's easier to
compare the visual positions, so I've gone with that for now. I don't
imagine it would be a huge speedup to do less coordinate conversions for
this, so it's unlikely I'll do anything else here for a good while.