Only eat mouseup for actions that run successfully
I don't remember why this was in here. It's in my original prototype for
the input manager but it doesn't seem immediately necessary for
anything, looking again.

Either way, this was interfering with the piano's mouseup event
listener, so I'm removing this for now until I investigate input
handling again in detail (which I've been putting off until I look at
replacing dockview with something else).