Only set drag bit when clicking inside a panel
This makes it so that e.g. after moving the timeline panel around, it's
still possible to use the mouse wheel for zooming. Previously you'd have
to click inside the timeline first, so it could get a mouseup event and
the input manager could clear the drag bit.

There was another issue where I registered the dockview panel element as
the area element for the timeline, but that was too broad, as it
includes the tab element by way of the search for an appropriate parent
in the input manager.

Something about this logic still feels off to me, like I should be
tracking a drag bit per mouse button, or something like that...