Add low quality time stretching for samples
Performance is not great and will have to be addressed if this is kept.

What was really stopping me from adding this though, was that the timing
was very off, compared with no time stretching (resampling) and a
similar algorithm (I presume) present in REAPER ("simple windowed"), but
it seems okay now.

The current commands are just so that this can be tested in some way.
Ideally there should be a dialog for these, and probably an interactive
way to set the pitch shift factor.